feat: graphql service
