+++ .gitignore
@@ -1,5 +1,4 @@
 # Logs
-### hello test
 logs
 *.log
 npm-debug.log*
