+++ package.json
@@ -73,6 +73,7 @@
     "@types/path-to-regexp": "^1.7.0",
     "@types/sinon": "^7.5.1",
     "@types/xml-js": "^1.0.0",
+    "apollo-server-lambda": "^2.9.15",
     "aws-lambda": "^0.1.2",
     "aws-sdk": "^2.590.0",
     "axios": "^0.19.0",
