+++ src/policies/authentification/ApiKey.ts
@@ -1,5 +1,4 @@
 import { get } from 'lodash';
-import { Context } from "../../router";
 import { tryExecuteFieldValue } from "../index";
 import Policy, { ExecutionContext, PolicyCategory, Scope, YapPolicy } from '../policy';
